# Backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Job Aggregator Platform API
  description: API documentation for the Job Aggregator Platform
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://api.jobaggregator.com/api
    description: Production server
paths:
  /auth/register:
    post:
      summary: Register a new user
      description: Register a new user with the provided details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict, user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      summary: Login a user
      description: Login a user with the provided credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: User logged in successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Unauthorized, invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      summary: Get current user details
      description: Get details of the currently logged-in user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized, invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /jobs:
    get:
      summary: Get all job listings
      description: Retrieve a list of all job listings
      responses:
        '200':
          description: List of job listings retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobListing'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new job listing
      description: Create a new job listing with the provided details
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreation'
      responses:
        '201':
          description: Job listing created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized, invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /jobs/{id}:
    get:
      summary: Get a specific job listing
      description: Retrieve details of a specific job listing
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the job listing
          schema:
            type: integer
      responses:
        '200':
          description: Job listing details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '404':
          description: Job listing not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update a job listing
      description: Update details of a specific job listing
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the job listing
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreation'
      responses:
        '200':
          description: Job listing updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized, invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Job listing not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a job listing
      description: Delete a specific job listing
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the job listing
          schema:
            type: integer
      responses:
        '200':
          description: Job listing deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized, invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Job listing not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /applications:
    get:
      summary: Get all applications
      description: Retrieve a list of all applications
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of applications retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Application'
        '401':
          description: Unauthorized, invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new application
      description: Create a new application with the provided details
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationCreation'
      responses:
        '201':
          description: Application created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized, invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /applications/{id}:
    get:
      summary: Get a specific application
      description: Retrieve details of a specific application
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the application
          schema:
            type: integer
      responses:
        '200':
          description: Application details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '401':
          description: Unauthorized, invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Application not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    UserRegistration:
      type: object
      properties:
        name:
          type: string
          example: Doe
        firstname:
          type: string
          example: John
        email:
          type: string
          format: email
          example: john@example.com
        password:
          type: string
          format: password
          example: mypassword123
        role:
          type: string
          enum: [user, recruiter, admin]
          example: user
        company:
          type: string
          example: Acme Corp
        siret:
          type: string
          example: "12345678900012"
      required:
        - name
        - firstname
        - email
        - password
        - role
    UserLogin:
      type: object
      properties:
        email:
          type: string
          format: email
          example: john@example.com
        password:
          type: string
          format: password
          example: mypassword123
      required:
        - email
        - password
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Doe
        firstname:
          type: string
          example: John
        email:
          type: string
          format: email
          example: john@example.com
        role:
          type: string
          enum: [user, recruiter, admin]
          example: user
        company:
          type: string
          example: Acme Corp
        siret:
          type: string
          example: "12345678900012"
        createdAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00.000Z"
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          $ref: '#/components/schemas/UserResponse'
    JobListing:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Software Engineer
        company:
          type: string
          example: Acme Corp
        location:
          type: string
          example: Paris, France
        contractType:
          type: string
          example: Full-time
        description:
          type: string
          example: We are looking for a software engineer to join our team.
        salaryMin:
          type: integer
          example: 50000
        salaryMax:
          type: integer
          example: 70000
        currency:
          type: string
          example: EUR
        experienceYears:
          type: integer
          example: 2
        remoteType:
          type: string
          example: Hybrid
        skills:
          type: array
          items:
            type: string
          example: ["JavaScript", "React", "Node.js"]
        createdAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00.000Z"
    JobCreation:
      type: object
      properties:
        title:
          type: string
          example: Software Engineer
        company:
          type: string
          example: Acme Corp
        location:
          type: string
          example: Paris, France
        contractType:
          type: string
          example: Full-time
        description:
          type: string
          example: We are looking for a software engineer to join our team.
        salaryMin:
          type: integer
          example: 50000
        salaryMax:
          type: integer
          example: 70000
        currency:
          type: string
          example: EUR
        experienceYears:
          type: integer
          example: 2
        remoteType:
          type: string
          example: Hybrid
        skills:
          type: array
          items:
            type: string
          example: ["JavaScript", "React", "Node.js"]
      required:
        - title
        - company
        - location
        - contractType
        - description
        - salaryMin
        - salaryMax
        - currency
        - experienceYears
        - remoteType
        - skills
    JobResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/JobListing'
    Application:
      type: object
      properties:
        id:
          type: integer
          example: 1
        jobId:
          type: integer
          example: 1
        userId:
          type: integer
          example: 1
        cvPath:
          type: string
          example: /uploads/cvs/cv.pdf
        coverLetterPath:
          type: string
          example: /uploads/cover_letters/cover_letter.pdf
        extraDocumentsPath:
          type: string
          example: /uploads/extra_documents/extra_document.pdf
        applicantFirstname:
          type: string
          example: John
        applicantName:
          type: string
          example: Doe
        applicantGender:
          type: string
          example: Male
        applicantEmail:
          type: string
          format: email
          example: john@example.com
        applicantPhone:
          type: string
          example: "1234567890"
        applicantPostalCode:
          type: string
          example: "75001"
        applicantCity:
          type: string
          example: Paris
        appliedAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00.000Z"
    ApplicationCreation:
      type: object
      properties:
        jobId:
          type: integer
          example: 1
        cvPath:
          type: string
          example: /uploads/cvs/cv.pdf
        coverLetterPath:
          type: string
          example: /uploads/cover_letters/cover_letter.pdf
        extraDocumentsPath:
          type: string
          example: /uploads/extra_documents/extra_document.pdf
        applicantFirstname:
          type: string
          example: John
        applicantName:
          type: string
          example: Doe
        applicantGender:
          type: string
          example: Male
        applicantEmail:
          type: string
          format: email
          example: john@example.com
        applicantPhone:
          type: string
          example: "1234567890"
        applicantPostalCode:
          type: string
          example: "75001"
        applicantCity:
          type: string
          example: Paris
      required:
        - jobId
        - applicantFirstname
        - applicantName
        - applicantEmail
        - applicantPhone
        - applicantPostalCode
        - applicantCity
    ApplicationResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Application'
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation successful
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
              example: Error message
            status:
              type: integer
              example: 400
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT